[unbuildable]
- Configuration Gfx got new attribute - CODE which will tell code which texture to take
- Replacing SDL with SFML
- Every object(item and entity for now) will have InitGfx() and its own sprite ... or default if not found
  (test implementation is in gem object)
- Removed theme class - redundant

+++ Views/Base/BaseView.cs
@@ -11,9 +11,9 @@ namespace Castles
     public class BaseView : IView
     {
         //protected Point origin { get; set; }
-        protected Vector2f origin { get; set; }
+        protected Vector2i origin { get; set; }
 
-        public BaseView(Vector2f origin)
+        public BaseView(Vector2i origin)
         {
             this.origin = origin;
         }
@@ -28,9 +28,9 @@ namespace Castles
         /// <param name="X">X</param>
         /// <param name="Y">Y</param>
         /// <returns>Shifted vector</returns>
-        protected Vector2f OP(int X, int Y)
+        protected Vector2i OP(int X, int Y)
         {
-            return new Vector2f(X+origin.X, Y+origin.Y);
+            return new Vector2i(X+origin.X, Y+origin.Y);
         }
     }
 }
